+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jfieux <jfieux@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/29 10:54:59 by jfieux            #+#    #+#             */
-/*   Updated: 2021/03/12 11:37:38 by jfieux           ###   ########.fr       */
+/*   Updated: 2021/03/12 11:55:16 by jfieux           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,8 +61,8 @@ char			*ft_malres(long long int num, char *res, int *i);
 char			*ft_umalres(unsigned long long int num, char *res, int *i);
 char			*ft_init_int(long long int num, char *res, int i);
 char			*ft_init_uint(unsigned long long int num, char *res, int i);
-int				ft_treat_hexa(long int nb, char **arg, int maj);
-char			*ft_init_int_hexa(long int num, char *res, int maj, int i);
+int				ft_treat_hexa(unsigned long long int nb, char **arg, int maj);
+char			*ft_init_int_hexa(unsigned long long int num, char *res, int maj, int i);
 char			*ft_malloc_tmp(t_struct *info, char *arg);
 int				ft_nb_space(char *flag, t_struct *info, int s, va_list param);
 int				ft_atoi(char *str);
